This is the version that is ready for animation
feel free to edit it if you want, esp. the laucher pole is a bit
thick…. but when animating just the end should move, then of course the
legs, so the launcher pole end and the legs are all that should be
animated, unless there is something else you see fit to animate.
Thanks,
The Great Vladimir Skovzky.